Trabajo sombre las sombras (Fixes #146) + Nueva propiedad ShadowSprite._todas_las sombras: una lista con todas   las sombras creadas para ese objeto, de manera de poder correr   reubicar() sobre todas las sombras (incluso las que no se ven) y así   evitar el desfasaje. + El Sprite Sombra ahora tiene un parámetro "dirección" para mostrar en   el debugging. * El parámetro "n" de add_shadow() fue cambiado a "direccion" para   clarificar lo que realmente es.
Otros cambios
* Camara.panear(): sign_x y sign_y ahora son ints en lugar de floats.
* QuickCircularMenu: todos los elementos de cualquiera de las cascadas
  tienen un indes, no solo la inicial. También (creo) corregido el tema
  de recordar la posición del menú, independientemente de cual sea la
  cascada.

+++ engine/globs/renderer.py
@@ -178,9 +178,9 @@ class Camara:
         sign_x, sign_y = 0, 0
         # Extraer el signo + o -
         if dx:
-            sign_x = dx / abs_x  # +1 ó -1
+            sign_x = int(dx / abs_x)  # +1 ó -1
         if dy:
-            sign_y = dy / abs_y  # +1 ó -1
+            sign_y = int(dy / abs_y)  # +1 ó -1
 
         while abs_x:
             # acá restauramos el valor para hacer la comparación
